split eslint rules
